// frontend/src/app/models/reservation.model.ts
export interface Reservation {
  _id?: string;               // Optional field for the reservation ID (generated by MongoDB)
  userId: string; 
  fullName: string;          // User's full name
  facilityId: string;         // ID of the reserved facility
  facilityName?: string;      // Name of the reserved facility (optional)
  price: number;              // Price of the reservation
  address: string;
  contactNumber: string;            // User's address
  reservationDate: Date;      // Date of the reservation
  startTime: string;          // Start time of the reservation (e.g., "10:00")
  endTime: string;            // End time of the reservation (e.g., "12:00")
  purpose: string;            // Purpose of the reservation
  status?: string;            // Status of the reservation (optional, e.g., "Pending", "Confirmed")
}

export interface TimeSlot {
  start: string;
  end: string;
}

export interface AvailableDate {
  day: string;
  slots: TimeSlot[];
}